+++ lib/theme/color.dart
@@ -5,4 +5,5 @@ class color {
   static const mainGreen = Color(0xFF9CA986);
   static const darkGreen = Color(0xFF808D7C);
   static const darkestGreen = Color(0xFF5F6F65);
+  static const white = Colors.white;
 }
